Enum sage_lisp::Expr

pub enum Expr {

    None,
    Float(f64),
    Int(i64),
    String(String),
    Symbol(Symbol),
    Bool(bool),
    List(Vec<Expr>),
    Tree(BTreeMap<Expr, Expr>),
    Map(HashMap<Expr, Expr>),
    Many(Arc<Vec<Expr>>),
    Quote(Box<Expr>),
    Err(Box<Self>),
    Function(Option<Box<Env>>, Vec<Expr>, Box<Expr>),
    Builtin(Builtin),

}

A lisp expression to be evaluated

Variants

None

The unit value (nil)

Float(f64)

A floating point number

Int(i64)

A signed 64-bit integer

String(String)

A string

Symbol(Symbol)

A symbol

Bool(bool)

A boolean

List(Vec<Expr>)

A list of expressions.

When evaluated, this is used to represent function calls, where the first element is the function to call, and the rest of the elements are the arguments.

When used as a data structure, this is used to represent a list of values. It can be indexed by position, and can be used to store a sequence of values.

This is stored as a vector, not a linked list, so that it can be indexed efficiently.

Tree(BTreeMap<Expr, Expr>)

An ordered map of expressions.

This is helpful when the user desires the ordered properties of a BTreeMap, such as ordering by key, but with worse time complexities than a HashMap.

Map(HashMap<Expr, Expr>)

A map of expressions.

This is helpful when the user wants the time complexities associated with a HashMap, such as O(1) for insertion, deletion, and lookup, but with no ordering.

Many(Arc<Vec<Expr>>)

A block of expressions to be evaluated in order.

This is used to group multiple expressions together, and to evaluate them in sequence. The result of the block is the result of the last expression in the block.

This is useful for defining functions, where you want to evaluate multiple expressions in order, and return the result of the last expression as the result of the function.

Quote(Box<Expr>)

A quoted expression.

This allows for lazy evaluation of expressions: when a quoted expression is evaluated, it returns the expression itself, without evaluating it. This is useful for defining special forms, or for returning unevaluated expressions from functions, like symbols.

Err(Box<Self>)

An error.

When an error occurs during evaluation, this is used to wrap an error value that is propagated up the call stack. This allows for error handling in the interpreter.

Function(Option<Box<Env>>, Vec<Expr>, Box<Expr>)

A function closure.

This is used to represent a function that takes arguments and returns a value. The function is defined by a list of arguments, and a body expression to evaluate.

Internally, the function also keeps track of the environment in which it was defined, which allows it to capture bindings to variables defined outside the function.

Builtin(Builtin)

A builtin function.

This is used to represent a function that is defined in Rust, and can be called from lisp.

Implementations

impl Expr

pub fn serialize<T: Serialize>(x: T) -> Self

Serialize a value into a Lisp expression.

pub fn deserialize<T: DeserializeOwned>(x: &Self) -> Result<T, Error>

Deserialize a Lisp expression into a value.

pub fn symbol(name: impl ToString) -> Self

Create a symbol Lisp expression from a string.

pub fn error(message: impl Into<Self>) -> Self

Wrap another expression in an error value.

This is useful for propagating errors up the call stack, and for handling errors in the interpreter.

pub fn quote(&self) -> Self

Quote an expression to prevent it from being evaluated.

pub fn apply(&self, args: &[Self]) -> Self

Apply a callable value to a list of arguments.

pub fn parse(input: &str) -> Result<Expr, String>

Parse a string into a Lisp expression.

If the string is a valid Lisp expression, it will return the parsed expression. If the string is not a valid Lisp expression, it will return an error message.